# server/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register/check-availability": {
      "post": {
        "operationId": "AuthController_checkAvailability",
        "summary": "检查注册字段可用性",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAvailabilityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "检查成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationAvailabilityResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/register/send-code": {
      "post": {
        "operationId": "AuthController_sendRegistrationCode",
        "summary": "发送注册验证码",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendRegistrationCodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "验证码已发送",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationCodeSentDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/register/verify-code": {
      "post": {
        "operationId": "AuthController_verifyRegistrationCode",
        "summary": "校验注册验证码",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRegistrationCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "验证码校验成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationVerificationResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "注册账号",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "账号登录",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "summary": "获取当前登录用户信息",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/admin/me": {
      "get": {
        "operationId": "AuthController_getAdminMe",
        "summary": "获取当前管理员信息",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/auth/github": {
      "get": {
        "operationId": "AuthController_githubLogin",
        "summary": "发起 GitHub OAuth 登录",
        "parameters": [
          {
            "name": "redirect",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/github/callback": {
      "get": {
        "operationId": "AuthController_githubCallback",
        "summary": "处理 GitHub OAuth 回调",
        "parameters": [
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/google": {
      "get": {
        "operationId": "AuthController_googleLogin",
        "summary": "发起 Google OAuth 登录",
        "parameters": [
          {
            "name": "redirect",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/google/callback": {
      "get": {
        "operationId": "AuthController_googleCallback",
        "summary": "处理 Google OAuth 回调",
        "parameters": [
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Blog System",
    "description": "Personal Blog System API - 自动生成的后端接口文档",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
      "RegisterAvailabilityDto": {
        "type": "object",
        "properties": {
          "registerType": {
            "type": "string",
            "description": "注册方式",
            "enum": [
              "email",
              "phone"
            ],
            "example": "email"
          },
          "contact": {
            "type": "string",
            "description": "待检查的联系地址（邮箱或手机号）",
            "example": "coder@example.com"
          },
          "username": {
            "type": "string",
            "description": "待检查的用户名",
            "example": "coder_su"
          },
          "nickname": {
            "type": "string",
            "description": "待检查的昵称",
            "example": "苏同学"
          }
        }
      },
      "RegistrationAvailabilityFieldDto": {
        "type": "object",
        "properties": {
          "available": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "object",
            "example": "用户名可用",
            "nullable": true
          },
          "normalizedValue": {
            "type": "object",
            "example": "coder_su",
            "nullable": true
          }
        },
        "required": [
          "available"
        ]
      },
      "RegistrationAvailabilityResponseDto": {
        "type": "object",
        "properties": {
          "contact": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RegistrationAvailabilityFieldDto"
              }
            ]
          },
          "username": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RegistrationAvailabilityFieldDto"
              }
            ]
          },
          "nickname": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RegistrationAvailabilityFieldDto"
              }
            ]
          }
        }
      },
      "SendRegistrationCodeDto": {
        "type": "object",
        "properties": {
          "registerType": {
            "type": "string",
            "description": "注册方式",
            "enum": [
              "email",
              "phone"
            ],
            "example": "email"
          },
          "contact": {
            "type": "string",
            "description": "邮箱地址或手机号",
            "example": "coder@example.com"
          }
        },
        "required": [
          "registerType",
          "contact"
        ]
      },
      "RegistrationCodeSentDto": {
        "type": "object",
        "properties": {
          "registerType": {
            "type": "string",
            "enum": [
              "email",
              "phone"
            ],
            "example": "email"
          },
          "maskedContact": {
            "type": "string",
            "example": "co***@example.com"
          },
          "expiresInSeconds": {
            "type": "number",
            "example": 600
          },
          "cooldownSeconds": {
            "type": "number",
            "example": 60
          },
          "debugCode": {
            "type": "object",
            "example": "238901",
            "nullable": true
          }
        },
        "required": [
          "registerType",
          "maskedContact",
          "expiresInSeconds",
          "cooldownSeconds"
        ]
      },
      "VerifyRegistrationCodeDto": {
        "type": "object",
        "properties": {
          "registerType": {
            "type": "string",
            "description": "注册方式",
            "enum": [
              "email",
              "phone"
            ],
            "example": "email"
          },
          "contact": {
            "type": "string",
            "description": "邮箱地址或手机号",
            "example": "coder@example.com"
          },
          "code": {
            "type": "string",
            "description": "6 位验证码",
            "example": "238901"
          }
        },
        "required": [
          "registerType",
          "contact",
          "code"
        ]
      },
      "RegistrationVerificationResponseDto": {
        "type": "object",
        "properties": {
          "registerType": {
            "type": "string",
            "enum": [
              "email",
              "phone"
            ],
            "example": "email"
          },
          "maskedContact": {
            "type": "string",
            "example": "co***@example.com"
          },
          "verificationToken": {
            "type": "string",
            "description": "验证码校验成功后用于最终注册的令牌",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.registration"
          },
          "expiresIn": {
            "type": "string",
            "example": "30m"
          }
        },
        "required": [
          "registerType",
          "maskedContact",
          "verificationToken",
          "expiresIn"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "registerType": {
            "type": "string",
            "description": "注册方式",
            "enum": [
              "email",
              "phone"
            ],
            "example": "email"
          },
          "verificationToken": {
            "type": "string",
            "description": "验证码校验成功后返回的注册令牌",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.registration"
          },
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "coder_su"
          },
          "password": {
            "type": "string",
            "description": "登录密码，必须包含字母和数字",
            "example": "Pass123456"
          },
          "nickname": {
            "type": "string",
            "description": "昵称",
            "example": "苏同学"
          }
        },
        "required": [
          "registerType",
          "verificationToken",
          "username",
          "password"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "0f4c5af2-1111-4444-8888-123456789abc"
          },
          "username": {
            "type": "string",
            "example": "author"
          },
          "email": {
            "type": "object",
            "example": "author@example.com",
            "nullable": true
          },
          "phone": {
            "type": "object",
            "example": "+8613800138000",
            "nullable": true
          },
          "nickname": {
            "type": "object",
            "example": "博主",
            "nullable": true
          },
          "avatar": {
            "type": "object",
            "example": "https://example.com/avatar.png",
            "nullable": true
          },
          "bio": {
            "type": "object",
            "example": "持续写作与构建产品。",
            "nullable": true
          },
          "registrationType": {
            "type": "string",
            "example": "email",
            "enum": [
              "email",
              "phone",
              "oauth"
            ]
          },
          "emailVerified": {
            "type": "boolean",
            "example": true
          },
          "phoneVerified": {
            "type": "boolean",
            "example": false
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "role": {
            "type": "object",
            "example": "author"
          },
          "lastLoginAt": {
            "type": "object",
            "example": "2026-04-19T10:00:00.000Z",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-04-19T10:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-04-19T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "phone",
          "nickname",
          "avatar",
          "bio",
          "registrationType",
          "emailVerified",
          "phoneVerified",
          "isActive",
          "role",
          "lastLoginAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "tokenType": {
            "type": "string",
            "example": "Bearer"
          },
          "expiresIn": {
            "type": "string",
            "example": "7d"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          }
        },
        "required": [
          "accessToken",
          "tokenType",
          "expiresIn",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "account": {
            "type": "string",
            "description": "用户名、邮箱或手机号",
            "example": "admin_manager"
          },
          "password": {
            "type": "string",
            "description": "登录密码",
            "example": "admin12345678"
          }
        },
        "required": [
          "account",
          "password"
        ]
      },
...
    }
  }
}